requirements.txt
# Core dependencies for the FastAPI management service (excluding yolov9-specific packages)
fastapi==0.116.1
GPUtil==1.4.0
nvidia-ml-py3==7.352.0
pandas==2.3.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
python-multipart==0.0.20
PyYAML==6.0.2
uvicorn==0.35.0
